requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
llama-cpp-python==0.3.19
